Refactor file structure and add Demo User button
