+++ frontend/src/App.tsx
@@ -1,8 +1,8 @@
 import "./App.css"
 import "./index.css"
 import { BrowserRouter, Routes, Route } from "react-router-dom"
-import Home from "./pages/Home"
-import NotFound from "./pages/NotFound"
+import Home from "./app/Home"
+import NotFound from "./app/NotFound"
 import PrivateRoute from "./components/PrivateRoute"
 import LoginPage from "@/app/login/page"
